#698 after rebase fixes
