+++ opba-protocols/xs2a-protocol-tests/xs2a-bdd-tests-common/src/main/java/de/adorsys/opba/protocol/xs2a/tests/e2e/stages/AccountInformationRequestCommon.java
@@ -3,14 +3,12 @@ package de.adorsys.opba.protocol.xs2a.tests.e2e.stages;
 import com.google.common.collect.ImmutableMap;
 import com.tngtech.jgiven.integration.spring.JGivenStage;
 import de.adorsys.opba.api.security.external.domain.OperationType;
-import de.adorsys.opba.api.security.external.service.RequestSigningService;
 import de.adorsys.opba.consentapi.model.generated.AuthViolation;
-import de.adorsys.opba.consentapi.model.generated.InlineResponse200;
-import de.adorsys.opba.consentapi.model.generated.ScaUserData;
 import de.adorsys.opba.protocol.xs2a.tests.GetTransactionsQueryParams;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -45,9 +43,6 @@ import static org.springframework.http.HttpHeaders.LOCATION;
 @JGivenStage
 @SuppressWarnings("checkstyle:MethodName") // Jgiven prettifies snake-case names not camelCase
 public class AccountInformationRequestCommon<SELF extends AccountInformationRequestCommon<SELF>> extends RequestCommon<SELF> {
-
-    @ScenarioState
-    private List<AuthViolation> violations;
 
     public SELF fintech_calls_list_accounts_for_anton_brueckner() {
         ExtractableResponse<Response> response = withAccountsHeaders(ANTON_BRUECKNER, requestSigningService, OperationType.AIS)
@@ -423,186 +418,6 @@ public class AccountInformationRequestCommon<SELF extends AccountInformationRequ
                                                     .captionMessage("{no.ctx.psuIpPort}");
         assertThat(this.violations).doesNotContain(authViolationIpPort);
 
-        return self();
-    }
-
-    private void startInitialInternalConsentAuthorization(String uriPath, String resource) {
-        ExtractableResponse<Response> response =
-                startInitialInternalConsentAuthorization(uriPath, resource, HttpStatus.ACCEPTED);
-        updateServiceSessionId(response);
-        updateRedirectCode(response);
-    }
-
-    private ExtractableResponse<Response> max_musterman_provides_sca_challenge_result() {
-        return provideParametersToBankingProtocolWithBody(
-                AUTHORIZE_CONSENT_ENDPOINT,
-                readResource("restrecord/tpp-ui-input/params/max-musterman-sca-challenge-result.json"),
-                HttpStatus.ACCEPTED
-        );
-    }
-
-    private void max_musterman_provides_password() {
-        startInitialInternalConsentAuthorization(
-                AUTHORIZE_CONSENT_ENDPOINT,
-                "restrecord/tpp-ui-input/params/max-musterman-password.json"
-        );
-    }
-
-
-    private ExtractableResponse<Response> startInitialInternalConsentAuthorization(String uriPath, String resource, HttpStatus status) {
-        return provideParametersToBankingProtocolWithBody(uriPath, readResource(resource), status);
-    }
-
-    private ExtractableResponse<Response> provideParametersToBankingProtocolWithBody(String uriPath, String body, HttpStatus status) {
-        ExtractableResponse<Response> response = RestAssured
-                .given()
-                    .header(X_REQUEST_ID, UUID.randomUUID().toString())
-                    .header(X_XSRF_TOKEN, UUID.randomUUID().toString())
-                    .cookie(AUTHORIZATION_SESSION_KEY, authSessionCookie)
-                    .queryParam(REDIRECT_CODE_QUERY, redirectCode)
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .body(body)
-                .when()
-                    .post(uriPath, serviceSessionId)
-                .then()
-                    .statusCode(status.value())
-                .extract();
-
-        this.responseContent = response.body().asString();
-        this.redirectUriToGetUserParams = response.header(LOCATION);
-        updateRedirectCode(response);
-        return response;
-    }
-
-    protected void updateNextConsentAuthorizationUrl(ExtractableResponse<Response> response) {
-        this.redirectUriToGetUserParams = response.header(LOCATION);
-    }
-
-    protected void updateServiceSessionId(ExtractableResponse<Response> response) {
-        this.serviceSessionId = response.header(SERVICE_SESSION_ID);
-    }
-
-    protected void updateRedirectCode(ExtractableResponse<Response> response) {
-        this.redirectCode = response.header(REDIRECT_CODE);
-    }
-
-    @SneakyThrows
-    private void updateAvailableScas() {
-        ExtractableResponse<Response> response = provideGetConsentAuthStateRequest();
-        InlineResponse200 parsedValue = new ObjectMapper()
-                                                .readValue(response.body().asString(), InlineResponse200.class);
-
-        this.availableScas = parsedValue.getConsentAuth().getScaMethods();
-        updateRedirectCode(response);
-    }
-
-    @SneakyThrows
-    private void readViolations() {
-        ExtractableResponse<Response> response = provideGetConsentAuthStateRequest();
-        InlineResponse200 parsedValue = new ObjectMapper()
-                                                 .readValue(response.body().asString(), InlineResponse200.class);
-
-        this.violations = parsedValue.getConsentAuth().getViolations();
-        updateRedirectCode(response);
-    }
-
-    private  ExtractableResponse<Response> provideGetConsentAuthStateRequest() {
-         return RestAssured
-                     .given()
-                            .header(X_REQUEST_ID, UUID.randomUUID().toString())
-                            .header(X_XSRF_TOKEN, UUID.randomUUID().toString())
-                            .cookie(AUTHORIZATION_SESSION_KEY, authSessionCookie)
-                            .queryParam(REDIRECT_CODE_QUERY, redirectCode)
-                     .when()
-                            .get(GET_CONSENT_AUTH_STATE, serviceSessionId)
-                     .then()
-                            .statusCode(HttpStatus.OK.value())
-                     .extract();
-    }
-
-    private String selectedScaBody(String scaName) {
-        return String.format(
-                "{\"scaAuthenticationData\":{\"SCA_CHALLENGE_ID\":\"%s\"}}",
-                this.availableScas.stream().filter(it -> it.getMethodValue().equals(scaName)).findFirst().get().getId()
-        );
-    }
-
-    public SELF fintech_calls_initiate_payment_for_anton_brueckner() {
-        String body = readResource("restrecord/tpp-ui-input/params/anton-brueckner-initiate-payment-body.json");
-        ExtractableResponse<Response> response = withAccountsHeaders(ANTON_BRUECKNER, requestSigningService, OperationType.PIS)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(body)
-        .when()
-            .post(PIS_SINGLE_PAYMENT_ENDPOINT, StandardPaymentProduct.SEPA_CREDIT_TRANSFERS.getSlug())
-        .then()
-            .statusCode(HttpStatus.ACCEPTED.value())
-            .extract();
-
-        updateServiceSessionId(response);
-        updateRedirectCode(response);
-        updateNextConsentAuthorizationUrl(response);
-        return self();
-    }
-
-    public SELF fintech_calls_initiate_payment_for_max_musterman() {
-        String body = readResource("restrecord/tpp-ui-input/params/max-musterman-initiate-payment-body.json");
-        ExtractableResponse<Response> response = withAccountsHeaders(MAX_MUSTERMAN, requestSigningService, OperationType.PIS)
-            .contentType(MediaType.APPLICATION_JSON_VALUE)
-            .body(body)
-        .when()
-            .post(PIS_SINGLE_PAYMENT_ENDPOINT, StandardPaymentProduct.SEPA_CREDIT_TRANSFERS.getSlug())
-        .then()
-            .statusCode(HttpStatus.ACCEPTED.value())
-            .extract();
-
-        updateServiceSessionId(response);
-        updateRedirectCode(response);
-        updateNextConsentAuthorizationUrl(response);
-        return self();
-    }
-
-    public SELF user_logged_in_into_opba_as_opba_user_with_credentials_using_fintech_supplied_url_pis(String username, String password) {
-        String fintechUserTempPassword = UriComponentsBuilder
-                .fromHttpUrl(redirectUriToGetUserParams).build()
-                .getQueryParams()
-                .getFirst(REDIRECT_CODE_QUERY);
-
-        ExtractableResponse<Response> response =  RestAssured
-                .given()
-                .header(X_REQUEST_ID, UUID.randomUUID().toString())
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .queryParam(REDIRECT_CODE_QUERY, fintechUserTempPassword)
-                .body(ImmutableMap.of(LOGIN, username, PASSWORD, password))
-                .when()
-                .post(PIS_LOGIN_USER_ENDPOINT, serviceSessionId)
-                .then()
-                .statusCode(HttpStatus.ACCEPTED.value())
-                .extract();
-
-        this.authSessionCookie = response.cookie(AUTHORIZATION_SESSION_KEY);
-        return self();
-    }
-
-    public SELF user_max_musterman_provided_sca_challenge_result_to_embedded_authorization_and_sees_redirect_to_fintech_ok_pis() {
-        assertThat(this.redirectUriToGetUserParams).contains("sca-result").doesNotContain("wrong=true");
-        ExtractableResponse<Response> response = max_musterman_provides_sca_challenge_result();
-        assertThat(response.header(LOCATION)).contains("pis").contains("consent-result");
-        return self();
-    }
-
-    public SELF user_anton_brueckner_sees_that_he_needs_to_be_redirected_to_aspsp_and_redirects_to_aspsp_pis() {
-        ExtractableResponse<Response> response = withDefaultHeaders(ANTON_BRUECKNER, requestSigningService, OperationType.PIS)
-                .cookie(AUTHORIZATION_SESSION_KEY, authSessionCookie)
-                .queryParam(REDIRECT_CODE_QUERY, redirectCode)
-                .when()
-                .get(GET_CONSENT_AUTH_STATE, serviceSessionId)
-                .then()
-                .statusCode(HttpStatus.OK.value())
-                .extract();
-
-        this.redirectUriToGetUserParams = response.header(LOCATION);
-        updateServiceSessionId(response);
-        updateRedirectCode(response);
         return self();
     }
 }
